Fix matchFile() to perform list.file.only() if the desired file is not in the working directory.
Fix error in setupSandbox() that caused it to only handle the first file.

Complete changeExtension() and correct its spelling.

Pull observeEvent(rubrics()) out from observeEvent(c(input$selectStudent, input$currentProblem)) so it only updates currentFiles().

Add to observeEvent(input$outerTabs) so that when the "Grading" tab is selected, currentFiles() and thisPath() are updated and the action buttons are correctly dis/enabled.